Use peek to implement print
The functionaliy of KeyValuePrinter is replaced with PrintAction which
implements ForeachAction and can be passed to KStreamPeek.
We therefore can get rid of KeyValuePrinter.
My only concern is that ForeachAction does not have access to the
ProcessorContext, which means, that the topic name cannot be used within
the deserialization process.
For now, the a null value is filled in as a placeholder for topic.
